+++ NEWS.md
@@ -12,6 +12,7 @@
 * Made "c" in "calendar" capital, added tooltip to "Generate Report" button, made "Disclaimer" bold, rephrased first sentence, and removed second (#77)
 * Changed "genre" to "gender" (#79)
 * In report: changed multiplication symbol * to \cdot, reviewed disclaimer, fixed typos, church types translated to English, fixed tables columns names (no underscore, leave space between words), date of birth in Parameters section changed in the form: dd-mm-yyyy (#82)
+* Fixed numeric input so that it doesn't crash when no value is inserted (#81)
 
 fixes #82